+++ cmd/influxd/run/server_test.go
@@ -5238,7 +5238,7 @@ func TestServer_Query_ShowQueries_Future(t *testing.T) {
 	}
 
 	writes := []string{
-		fmt.Sprintf(`cpu,host=server01 value=100 %d`, models.MaxNanoTime.UnixNano()),
+		fmt.Sprintf(`cpu,host=server01 value=100 %d`, models.MaxNanoTime),
 	}
 
 	test := NewTest("db0", "rp0")
@@ -6227,7 +6227,7 @@ func TestServer_Query_LargeTimestamp(t *testing.T) {
 	defer s.Close()
 
 	writes := []string{
-		fmt.Sprintf(`cpu value=100 %d`, models.MaxNanoTime.UnixNano()),
+		fmt.Sprintf(`cpu value=100 %d`, models.MaxNanoTime),
 	}
 
 	test := NewTest("db0", "rp0")
@@ -6238,8 +6238,8 @@ func TestServer_Query_LargeTimestamp(t *testing.T) {
 		&Query{
 			name:    `select value at max nano time`,
 			params:  url.Values{"db": []string{"db0"}},
-			command: fmt.Sprintf(`SELECT value FROM cpu WHERE time <= %d`, models.MaxNanoTime.UnixNano()),
-			exp:     `{"results":[{"series":[{"name":"cpu","columns":["time","value"],"values":[["` + models.MaxNanoTime.Format(time.RFC3339Nano) + `",100]]}]}]}`,
+			command: fmt.Sprintf(`SELECT value FROM cpu WHERE time <= %d`, models.MaxNanoTime),
+			exp:     `{"results":[{"series":[{"name":"cpu","columns":["time","value"],"values":[["` + time.Unix(0, models.MaxNanoTime).Format(time.RFC3339Nano) + `",100]]}]}]}`,
 		},
 	}...)